Make single_process_pg a pytest.fixture
