+++ tests/core/test_lightning_module.py
@@ -23,7 +23,6 @@ from pytorch_lightning import Trainer
 from pytorch_lightning.loggers import TensorBoardLogger
 from pytorch_lightning.utilities import _TORCH_SHARDED_TENSOR_AVAILABLE
 from tests.helpers import BoringModel
-from tests.helpers.process_group import single_process_pg
 from tests.helpers.runif import RunIf
 
 
@@ -315,27 +314,24 @@ class BoringModelWithShardedTensor(BoringModel):
 @pytest.mark.skipif(
     not _TORCH_SHARDED_TENSOR_AVAILABLE, reason="Test requires the torch version to support `ShardedTensor`"
 )
-def test_sharded_tensor_state_dict(tmpdir):
-    # Initialize the global process group since sharded tensor factory
-    # functions depend on a process group
-    with single_process_pg():
-        spec = dist._sharding_spec.ChunkShardingSpec(
-            dim=0,
-            placements=[
-                "rank:0/cpu",
-            ],
-        )
+def test_sharded_tensor_state_dict(tmpdir, single_process_pg):
+    spec = dist._sharding_spec.ChunkShardingSpec(
+        dim=0,
+        placements=[
+            "rank:0/cpu",
+        ],
+    )
 
-        m_0 = BoringModelWithShardedTensor(spec)
-        m_0.sharded_tensor.local_shards()[0].tensor.fill_(1)
-        assert "sharded_tensor" in m_0.state_dict(), 'Expect "sharded_tensor" to appear in the state dict'
+    m_0 = BoringModelWithShardedTensor(spec)
+    m_0.sharded_tensor.local_shards()[0].tensor.fill_(1)
+    assert "sharded_tensor" in m_0.state_dict(), 'Expect "sharded_tensor" to appear in the state dict'
 
-        m_1 = BoringModelWithShardedTensor(spec)
-        assert not torch.allclose(
-            m_1.sharded_tensor.local_shards()[0].tensor, m_0.sharded_tensor.local_shards()[0].tensor
-        ), "Expect the shards to be different before `m_1` loading `m_0`'s state dict"
+    m_1 = BoringModelWithShardedTensor(spec)
+    assert not torch.allclose(
+        m_1.sharded_tensor.local_shards()[0].tensor, m_0.sharded_tensor.local_shards()[0].tensor
+    ), "Expect the shards to be different before `m_1` loading `m_0`'s state dict"
 
-        m_1.load_state_dict(m_0.state_dict(), strict=False)
-        assert torch.allclose(
-            m_1.sharded_tensor.local_shards()[0].tensor, m_0.sharded_tensor.local_shards()[0].tensor
-        ), "Expect the shards to be same after `m_1` loading `m_0`'s state dict"
+    m_1.load_state_dict(m_0.state_dict(), strict=False)
+    assert torch.allclose(
+        m_1.sharded_tensor.local_shards()[0].tensor, m_0.sharded_tensor.local_shards()[0].tensor
+    ), "Expect the shards to be same after `m_1` loading `m_0`'s state dict"
